+++ src/cpac/helpers/__init__.py
@@ -1,5 +1,16 @@
 '''Hepler functions for cpac Python package.'''
 import re
+from itertools import chain
+
+TODOs = {'persisting_containers': 'Some Docker containers unexpectedly '
+                                  'persist after cpac finishes. To clear '
+                                  'them, run\n    '
+                                  r'1. `docker ps` to list the containers'
+                                  '\n  For each C-PAC conatainer that '
+                                  'persists, run\n    '
+                                  r'2. `docker attach <container_name>`'
+                                  '\n    '
+                                  r'3. `exit`'}
 
 
 def get_extra_arg_value(extra_args, argument):
@@ -28,15 +39,13 @@ def get_extra_arg_value(extra_args, argument):
     ...     '--participant_ndx 3'], 'participant_ndx')
     '3'
     '''
-    pattern = r'^\-*' + argument + r'([=\s]{1}.*)$'
+    extra_args = list(chain.from_iterable([
+        re.split(r'[=\s]', arg) for arg in extra_args]))
 
     for index, item in enumerate(extra_args):
-        if re.match(pattern, item) is not None:
-            for sep in {'=', ' '}:
-                if sep in item:
-                    return item.split(sep, 1)[1]
-            if len(extra_args) > index:
-                return extra_args[index + 1]
+        if item.startswith('-') and item.lstrip('-') == argument:
+            return extra_args[index + 1]
+    return None
 
 
-__all__ = ['get_extra_arg_value']
+__all__ = ['get_extra_arg_value', 'TODOs']