+++ tests/test_cpac_group.py
@@ -1,15 +1,17 @@
-import pytest
 import sys
 
 from unittest import mock
 
+import pytest
+
 from cpac.__main__ import run
-from CONSTANTS import args_before_after, set_commandline_args
+from .CONSTANTS import args_before_after, set_commandline_args
 
 
 @pytest.mark.parametrize('argsep', [' ', '='])
-def test_utils_help(argsep, capsys, platform=None, tag=None):
+def test_utils_help(argsep, capsys, platform, tag):
     def run_test(argv, platform):
+        argv = [arg for arg in argv if arg]
         with mock.patch.object(sys, 'argv', argv):
             run()
             captured = capsys.readouterr()